web/requirements.txt
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==8.0.1
cryptography==3.4.7
emoji-country-flag==1.2.4
Flask==2.0.1
google-api-core==1.28.0
google-auth==1.30.1
google-cloud-core==1.6.0
google-cloud-firestore==2.1.1
googleapis-common-protos==1.53.0
grpcio==1.38.0
idna==2.10
importlib-metadata==4.3.1
inetlab==0.1.3
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
msal==1.12.0
packaging==20.9
proto-plus==1.18.1
protobuf==3.17.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
rsa==4.7.2
six==1.16.0
typing-extensions==3.10.0.0
ua-parser==0.10.0
urllib3==1.26.5
user-agents==2.2.0
Werkzeug==2.0.1
zipp==3.4.1
